+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 0.0.152
-- [Update] update doc
+- [Update] Create flutter page with auto Route
 
 # 0.0.151
 - [Update] Renamed `PngImage`/`SvgIcon` enums to `PngAssets`/`SvgAssets` in